Bug fixes and cleanup, rename variables

+++ molalignlib/eigen-lapack.f90
@@ -24,37 +24,37 @@ end interface
 contains
 
 function leasteigval_sp(a) result(leasteigval)
-   real(r4), intent(inout) :: a(4, 4)
+   real(real32), intent(inout) :: a(4, 4)
    integer :: info
-   real(r4) :: leasteigval
-   real(r4) :: w(4), work(25)
+   real(real32) :: leasteigval
+   real(real32) :: w(4), work(25)
    call ssyev('N', 'U', 4, a, 4, w, work, 25, info)
    leasteigval = w(1)
 end function
 
 function leasteigval_dp(a) result(leasteigval)
-   real(r8), intent(inout) :: a(4, 4)
+   real(real64), intent(inout) :: a(4, 4)
    integer :: info
-   real(r8) :: leasteigval
-   real(r8) :: w(4), work(25)
+   real(real64) :: leasteigval
+   real(real64) :: w(4), work(25)
    call dsyev('N', 'U', 4, a, 4, w, work, 25, info)
    leasteigval = w(1)
 end function
 
 function leasteigvec_sp(a) result(leasteigvec)
-   real(r4), intent(inout) :: a(4, 4)
+   real(real32), intent(inout) :: a(4, 4)
    integer :: info
-   real(r4) :: w(4), work(25)
-   real(r4) :: leasteigvec(4)
+   real(real32) :: w(4), work(25)
+   real(real32) :: leasteigvec(4)
    call ssyev('V', 'U', 4, a, 4, w, work, 25, info)
    leasteigvec = a(:, 1)
 end function
 
 function leasteigvec_dp(a) result(leasteigvec)
-   real(r8), intent(inout) :: a(4, 4)
+   real(real64), intent(inout) :: a(4, 4)
    integer :: info
-   real(r8) :: leasteigvec(4)
-   real(r8) :: w(4), work(25)
+   real(real64) :: leasteigvec(4)
+   real(real64) :: w(4), work(25)
    call dsyev('V', 'U', 4, a, 4, w, work, 25, info)
    leasteigvec = a(:, 1)
 end function
